pyproject.toml
[project]
name = "github-gregl83"
version = "0.1.0"
description = "Github profile repository for gregl83"
requires-python = ">=3.10"
classifiers = [
    "Private :: Do Not Upload",
]
dependencies = [
    "google-genai>=1.52.0",
    "jinja2>=3.1.6",
    "pillow>=12.0.0",
    "xai-sdk>=1.4.1",
]

[project.optional-dependencies]
dev = []